+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.5.1] UNRELEASED
 ### Modified
-- Checksum can apply to `date`, `datetime`, `Mapping` or `Iterable` objects.
+- Checksum can apply to `date` and `datetime` objects.
 
 ## [0.5.0] 2026-03-05
 ### Modified